+++ lib/src/async_edit_builder.dart
@@ -2,20 +2,21 @@ import 'package:fetcher/fetcher.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 
-typedef DataEditWidgetBuilder<T> = Widget Function(BuildContext context, T value, ValueSetter<T> commit);
+typedef DataEditWidgetBuilder<T> = Widget Function(BuildContext context, T value, ValueSetter<T> submit);
 
-/// A widget that allow to fetch a value asynchronously, and then to run another task asynchronously (to commit new value for instance) while updating UI with new value.
+/// A widget that allow to fetch a value asynchronously, and then to run another task asynchronously (to submit new value for instance) while updating UI with new value.
 /// Handle all states (loading, errors, onSuccess).
 /// It's actually a mix of [FetchBuilder] and [SubmitBuilder] combined.
 /// Typically used for component that needs to fetch a value and then edit that value.
-/// Example : an async switch
+/// DON'T use this if fetch and submit tasks are not related (prefer using separate [FetchBuilder] and [SubmitBuilder]).
+/// Example : an async switch, that fetch the current value, and then can submit the new value.
 class AsyncEditBuilder<T> extends StatefulWidget {
   const AsyncEditBuilder({
     super.key,
     this.config,
     this.fetchingBuilder,
     required this.fetchTask,
-    required this.commitTask,
+    required this.submitTask,
     required this.builder,
     this.onEditSuccess,
   });
@@ -26,21 +27,21 @@ class AsyncEditBuilder<T> extends StatefulWidget {
 
   /// Widget to display while fetching
   /// Default to [config.fetchingBuilder]
-  /// If you want to change the committing widget, you should use [config.fetchingBuilder] instead.
+  /// If you want to change the submit widget, you should use [config.fetchingBuilder] instead.
   final WidgetBuilder? fetchingBuilder;
 
   /// Task that fetch value
   /// If task throws, it will be properly handled
   final AsyncValueGetter<T> fetchTask;
 
-  /// Task that commit modification
+  /// Task that submit modification
   /// If task throws, it will be properly handled
-  final AsyncValueSetter<T> commitTask;
+  final AsyncValueSetter<T> submitTask;
 
   /// Child widget builder
   final DataEditWidgetBuilder<T> builder;
 
-  /// Called after [commitTask] is successfully executed.
+  /// Called after [submitTask] is successfully executed.
   final AsyncValueSetter<T>? onEditSuccess;
 
   @override
@@ -72,7 +73,7 @@ class _AsyncEditBuilderState<T> extends State<AsyncEditBuilder<T>> {
             context,
             data,
             (data) => runTask(() async {
-              await widget.commitTask(data);
+              await widget.submitTask(data);
               return data;
             }),
           ),